ci: fix auth test type mismatch for roles fixture query
The `ensure_user_role` test helper queried `COALESCE(MAX(id_role), 0) + 1`
which PostgreSQL promotes to INT4, but the Rust code expected i16 (INT2).
Cast the result to smallint to match the DDL column type.

+++ kushim-auth/api/tests/auth.rs
@@ -58,7 +58,7 @@ async fn ensure_user_role(pool: &sqlx::PgPool) {
 
     let next_role_id = sqlx::query_scalar::<_, i16>(
         r#"
-        SELECT COALESCE(MAX(id_role), 0) + 1
+        SELECT (COALESCE(MAX(id_role), 0) + 1)::smallint
         FROM roles
         "#,
     )
